[ATIP] Move webdriver specfic steps from common to web
 And rename context dict from context.webs to context.apps to
 compatible with uiautomator
 BUG=https://crosswalk-project.org/jira/browse/XWALK-4650
