+++ tools/atip/atip/common/steps.py
@@ -30,42 +30,9 @@ import time
 from behave import step
 from atip.web import web
 
-
 @step(u'I wait for {timeout:d} seconds')
 def wait_for_timeout(context, timeout):
     time.sleep(timeout)
-
-
-@step(u'launch "{app_name}"')
-def launch_app_by_name(context, app_name):
-    web.launch_webapp_by_name(context, app_name)
-
-
-@step(u'I launch "{app_name}" with "{apk_pkg_name}" and "{apk_activity_name}"')
-def launch_app_by_names(context, app_name, apk_pkg_name, apk_activity_name):
-    web.launch_webapp_by_name(
-        context,
-        app_name,
-        apk_pkg_name,
-        apk_activity_name)
-
-
-@step(u'switch to "{app_name}"')
-def switch_to_app_name(context, app_name):
-    if app_name in context.webs:
-        context.web = context.webs[app_name]
-        assert True
-    else:
-        assert False
-
-
-@step(u'quit "{app_name}"')
-def quit_app_by_name(context, app_name):
-    if app_name in context.webs:
-        context.web.quit()
-        assert True
-    else:
-        assert False
 
 
 @step(
